backend/backend/requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
fastapi==0.75.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
mido==1.2.10
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
pandas==1.4.2
pretty-midi==0.2.9
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
